+++ src/handler/http/request/search/mod.rs
@@ -63,8 +63,8 @@ pub mod query_manager;
 pub mod saved_view;
 pub mod search_inspector;
 pub mod search_job;
-pub(crate) mod utils;
 pub mod search_stream;
+pub(crate) mod utils;
 
 async fn can_use_distinct_stream(
     org: &str,
@@ -861,8 +861,8 @@ pub async fn build_search_request_per_field(
     let mut requests = Vec::new();
     for field in fields {
         let sql = if no_count {
-            // we use min(0) as a hack to do streaming aggregation but actually return 0, essentially
-            // we are not counting the values
+            // we use min(0) as a hack to do streaming aggregation but actually return 0,
+            // essentially we are not counting the values
             format!(
                 "SELECT \"{field}\" AS zo_sql_key, min(0) AS zo_sql_num FROM \"{distinct_prefix}{stream_name}\" {sql_where} GROUP BY zo_sql_key"
             )
